move localizederror exception to break circdep
